config changes

+++ common/nengiConfig.js
@@ -24,7 +24,7 @@ const config = {
     UPDATE_RATE: 20, 
 
     ID_BINARY_TYPE: nengi.UInt16,
-    TYPE_BINARY_TYPE: nengi.UInt16, 
+    TYPE_BINARY_TYPE: nengi.UInt8, 
 
     ID_PROPERTY_NAME: 'nid',
     TYPE_PROPERTY_NAME: 'ntype',
